Create drawScreen()
Draw the contents of the Chip8 struct's screen memory to the ncurses
screen. Currently, this is a buggy, flickering mess. I'm not sure if
that's because I'm running at my computer's clock speed rather than a
slower rate, or if it's because of some issue with how I'm handling
ncurses. Either way, I'll fix it soon.